refactor: update order tracking and identification
- Switched from orderNumber to unique database ID for order referencing.
- Updated order date format to display full month names.
- Adjusted order sequence generation in Google Apps Script and corrected email template variables.

+++ src/components/CustomerDashboard.tsx
@@ -198,7 +198,7 @@ export const CustomerDashboard: React.FC = () => {
                       </span>
                     </div>
                     <div className="text-xs text-slate-400">
-                      Placed on {new Date(order.createdAt).toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })}
+                      Placed on {new Date(order.createdAt).toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })}
                     </div>
                   </div>
